fix(bios): cast bios data to 32 bit since we're reading a word

+++ src/bus.cpp
@@ -13,5 +13,5 @@ static u32 BIOSReadHandler(u32 address);
 u32 Bus::BIOSReadHandler(u32 address) {
     const u32 offset = address & UINT32_C(0x7FFFF);
 
-    return g_bios[offset];
+    return static_cast<u32>(g_bios[offset]);
 }
